Create LICENSE
Because of its simplicity and adaptability, we decided to employ the MIT License for our WebNotes project. With only a few limitations, the MIT License allows anybody to use, modify, and distribute our code without restriction. This is in line with our mission to promote community cooperation and creativity. The license is also a well-known and developer-friendly option for open-source projects since it limits liability and provides the team with legal protection.